// apps/storefront/prismicio-types.d.ts
export type PosterSliceOneUp = prismic.SharedSliceVariation<
    'oneUp',
    Simplify<PosterSliceOneUpPrimary>,
    Simplify<PosterSliceOneUpItem>
>;
// ...
export type PosterSliceTwoUp = prismic.SharedSliceVariation<
    'twoUp',
    Simplify<PosterSliceTwoUpPrimary>,
    Simplify<PosterSliceTwoUpItem>
>;

/**
 * Primary content in *Poster → Primary*
 */
export interface PosterSliceFeaturePrimary {
    /**
     * Desktop Image field in *Poster → Primary*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.primary.desktop_image
     * - **Documentation**: https://prismic.io/docs/field#image
     */
    desktop_image: prismic.ImageField<never>;

    /**
     * Mobile Image field in *Poster → Primary*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.primary.mobile_image
     * - **Documentation**: https://prismic.io/docs/field#image
     */
    mobile_image: prismic.ImageField<never>;

    /**
     * Product Title field in *Poster → Primary*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.primary.product_title
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    product_title: prismic.KeyTextField;

    /**
     * Description field in *Poster → Primary*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.primary.description
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    description: prismic.KeyTextField;

    /**
     * Background Colour field in *Poster → Primary*
     *
     * - **Field Type**: Color
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.primary.background_colour
     * - **Documentation**: https://prismic.io/docs/field#color
     */
    background_colour: prismic.ColorField;

    /**
     * Text Theme Desktop field in *Poster → Primary*
     *
     * - **Field Type**: Boolean
     * - **Placeholder**: *None*
     * - **Default Value**: true
     * - **API ID Path**: poster.primary.text_theme_desktop
     * - **Documentation**: https://prismic.io/docs/field#boolean
     */
    text_theme_desktop: prismic.BooleanField;

    /**
     * Text Theme Mobile field in *Poster → Primary*
     *
     * - **Field Type**: Boolean
     * - **Placeholder**: *None*
     * - **Default Value**: true
     * - **API ID Path**: poster.primary.text_theme_mobile
     * - **Documentation**: https://prismic.io/docs/field#boolean
     */
    text_theme_mobile: prismic.BooleanField;
}

/**
 * Primary content in *Poster → Items*
 */
export interface PosterSliceFeatureItem {
    /**
     * Button Label field in *Poster → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.items[].button_label
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    button_label: prismic.KeyTextField;

    /**
     * Button Link field in *Poster → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.items[].button_link
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    button_link: prismic.KeyTextField;

    /**
     * Button Variant field in *Poster → Items*
     *
     * - **Field Type**: Select
     * - **Placeholder**: *None*
     * - **Default Value**: Base
     * - **API ID Path**: poster.items[].button_variant
     * - **Documentation**: https://prismic.io/docs/field#select
     */
    button_variant: prismic.SelectField<
        'Base' | 'Secondary' | 'Text' | 'Ghost' | 'Quite',
        'filled'
    >;

    /**
     * Button Icon field in *Poster → Items*
     *
     * - **Field Type**: Select
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.items[].button_icon
     * - **Documentation**: https://prismic.io/docs/field#select
     */
    button_icon: prismic.SelectField<
        '-' | 'Northeast' | 'Right' | 'Left' | 'Globe' | 'Discover'
    >;
}
// ...
export interface PosterSliceFeatureStackItem {
    /**
     * Thumbnail field in *Poster → Items*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.items[].thumbnail
     * - **Documentation**: https://prismic.io/docs/field#image
     */
    thumbnail: prismic.ImageField<never>;

    /**
     * Desktop Image field in *Poster → Items*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.items[].desktop_image
     * - **Documentation**: https://prismic.io/docs/field#image
     */
    desktop_image: prismic.ImageField<never>;

    /**
     * Mobile Image field in *Poster → Items*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.items[].mobile_image
     * - **Documentation**: https://prismic.io/docs/field#image
     */
    mobile_image: prismic.ImageField<never>;

    /**
     * Description field in *Poster → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.items[].description
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    description: prismic.KeyTextField;

    /**
     * Background Colour field in *Poster → Items*
     *
     * - **Field Type**: Color
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.items[].background_colour
     * - **Documentation**: https://prismic.io/docs/field#color
     */
    background_colour: prismic.ColorField;
}
// ...
export interface PosterSliceFeaturesItem {
    /**
     * Button Variant field in *Poster → Items*
     *
     * - **Field Type**: Select
     * - **Placeholder**: *None*
     * - **Default Value**: Base
     * - **API ID Path**: poster.items[].button_variant
     * - **Documentation**: https://prismic.io/docs/field#select
     */
    button_variant: prismic.SelectField<
        'Base' | 'Secondary' | 'Text' | 'Ghost' | 'Quite',
        'filled'
    >;

    /**
     * Button Label field in *Poster → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.items[].button_label
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    button_label: prismic.KeyTextField;

    /**
     * Button Link field in *Poster → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.items[].button_link
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    button_link: prismic.KeyTextField;

    /**
     * Button Icon field in *Poster → Items*
     *
     * - **Field Type**: Select
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.items[].button_icon
     * - **Documentation**: https://prismic.io/docs/field#select
     */
    button_icon: prismic.SelectField<
        '-' | 'Northeast' | 'Right' | 'Left' | 'Globe' | 'Discover'
    >;
}

/**
 * Features variation for Poster Slice
 *
 * - **API ID**: `features`
 * - **Description**: Default
 * - **Documentation**: https://prismic.io/docs/slice
 */
export type PosterSliceFeatures = prismic.SharedSliceVariation<
    'features',
    Simplify<PosterSliceFeaturesPrimary>,
    Simplify<PosterSliceFeaturesItem>
>;

/**
 * Primary content in *Poster → Primary*
 */
export interface PosterSliceFeaturesAltPrimary {
    /**
     * Background Colour Left field in *Poster → Primary*
     *
     * - **Field Type**: Color
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.primary.background_colour_left
     * - **Documentation**: https://prismic.io/docs/field#color
     */
    background_colour_left: prismic.ColorField;

    /**
     * Desktop Image Left field in *Poster → Primary*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.primary.desktop_image_left
     * - **Documentation**: https://prismic.io/docs/field#image
     */
    desktop_image_left: prismic.ImageField<never>;

    /**
     * Mobile Image Left field in *Poster → Primary*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.primary.mobile_image_left
     * - **Documentation**: https://prismic.io/docs/field#image
     */
    mobile_image_left: prismic.ImageField<never>;

    /**
     * Background Colour Right field in *Poster → Primary*
     *
     * - **Field Type**: Color
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.primary.background_colour_right
     * - **Documentation**: https://prismic.io/docs/field#color
     */
    background_colour_right: prismic.ColorField;

    /**
     * Desktop Image Right field in *Poster → Primary*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.primary.desktop_image_right
     * - **Documentation**: https://prismic.io/docs/field#image
     */
    desktop_image_right: prismic.ImageField<never>;

    /**
     * Mobile Image Right field in *Poster → Primary*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.primary.mobile_image_right
     * - **Documentation**: https://prismic.io/docs/field#image
     */
    mobile_image_right: prismic.ImageField<never>;

    /**
     * Context field in *Poster → Primary*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.primary.context
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    context: prismic.KeyTextField;

    /**
     * Text Theme Desktop field in *Poster → Primary*
     *
     * - **Field Type**: Boolean
     * - **Placeholder**: *None*
     * - **Default Value**: true
     * - **API ID Path**: poster.primary.text_theme_desktop
     * - **Documentation**: https://prismic.io/docs/field#boolean
     */
    text_theme_desktop: prismic.BooleanField;

    /**
     * Text Theme Mobile field in *Poster → Primary*
     *
     * - **Field Type**: Boolean
     * - **Placeholder**: *None*
     * - **Default Value**: true
     * - **API ID Path**: poster.primary.text_theme_mobile
     * - **Documentation**: https://prismic.io/docs/field#boolean
     */
    text_theme_mobile: prismic.BooleanField;
}

/**
 * Primary content in *Poster → Items*
 */
export interface PosterSliceFeaturesAltItem {
    /**
     * Button Placement field in *Poster → Items*
     *
     * - **Field Type**: Select
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.items[].button_placement
     * - **Documentation**: https://prismic.io/docs/field#select
     */
    button_placement: prismic.SelectField<'Left' | 'Right'>;

    /**
     * Button Variant field in *Poster → Items*
     *
     * - **Field Type**: Select
     * - **Placeholder**: *None*
     * - **Default Value**: Base
     * - **API ID Path**: poster.items[].button_variant
     * - **Documentation**: https://prismic.io/docs/field#select
     */
    button_variant: prismic.SelectField<
        'Base' | 'Secondary' | 'Text' | 'Ghost' | 'Quite',
        'filled'
    >;

    /**
     * Button Label field in *Poster → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.items[].button_label
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    button_label: prismic.KeyTextField;

    /**
     * Button Link field in *Poster → Items*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.items[].button_link
     * - **Documentation**: https://prismic.io/docs/field#key-text
     */
    button_link: prismic.KeyTextField;

    /**
     * Button Icon field in *Poster → Items*
     *
     * - **Field Type**: Select
     * - **Placeholder**: *None*
     * - **API ID Path**: poster.items[].button_icon
     * - **Documentation**: https://prismic.io/docs/field#select
     */
    button_icon: prismic.SelectField<
        '-' | 'Northeast' | 'Right' | 'Left' | 'Globe' | 'Discover'
    >;
}

/**
 * Features Alt variation for Poster Slice
 *
 * - **API ID**: `featuresAlt`
 * - **Description**: Default
 * - **Documentation**: https://prismic.io/docs/slice
 */
export type PosterSliceFeaturesAlt = prismic.SharedSliceVariation<
    'featuresAlt',
    Simplify<PosterSliceFeaturesAltPrimary>,
    Simplify<PosterSliceFeaturesAltItem>
>;

/**
 * Slice variation for *Poster*
 */
type PosterSliceVariation =
    | PosterSliceOneUp
    | PosterSliceTwoUp
    | PosterSliceFeature
    | PosterSliceFeatureStack
    | PosterSliceFeatureStackVideo
    | PosterSliceFeatures
    | PosterSliceFeaturesAlt;

/**
 * Poster Shared Slice
 *
 * - **API ID**: `poster`
 * - **Description**: Poster
 * - **Documentation**: https://prismic.io/docs/slice
 */
export type PosterSlice = prismic.SharedSlice<'poster', PosterSliceVariation>;